EOC - Chapter 4 Webview
Completed project. WebView is loaded as default view directing to hackingWithSwift website. Page can be refreshed and there is a UIProgressBar in the UIToolbar also. All UIView have been manipulated via code without the use of StoryBoards in Interface builder.

There is a UIBarButton added to ope nan ActionSheet with the allowed sites the user can visit. KVO (Key Value Observer) pattern has been used in order to link the UIProgressBar to the estimatedProgress property on the webView. Further improvements to add in the future will include an initial TableView containing a list of the URLs you can visit. Also an UIAlertView explaining why a link cannot be clicked as it's not in the allowed list of sites.